+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/evaluator/ExpressionValue.java
@@ -40,7 +40,7 @@ class ExpressionValue implements LogicReadable {
                 case AstLiteralFloat n -> new ExpressionValue(processor, null, n.getDoubleValue());
                 case AstLiteralColor n -> new ExpressionValue(processor, null, n.getDoubleValue());
                 case AstLiteralString n -> new ExpressionValue(processor, n.getValue(), null);
-                case AstLiteralChar n -> new ExpressionValue(processor, n.getLongValue(), null);
+                case AstLiteralChar n -> new ExpressionValue(processor, null, n.getLongValue());
                 case AstIdentifier n -> processor.getMetadata().getIcons().isIconName(n.getName())
                         ? new ExpressionValue(processor, processor.getMetadata().getIcons().getIconValue(n.getName()).format(processor), null)
                         : new InvalidValue(processor);